+++ autogpt_platform/backend/backend/data/notifications.py
@@ -11,7 +11,6 @@ from prisma.types import UserNotificationBatchWhereInput
 # from backend.notifications.models import NotificationEvent
 from pydantic import BaseModel, EmailStr, Field, field_validator
 
-from backend.data.user import get_user_by_id, update_user_notification_preference
 from backend.server.v2.store.exceptions import DatabaseError
 
 from .db import transaction
@@ -404,39 +403,4 @@ async def get_user_notification_batch(
     except Exception as e:
         raise DatabaseError(
             f"Failed to get user notification batch for user {user_id} and type {notification_type}: {e}"
-        ) from e
-
-
-async def generate_unsubscribe_link(user_id: str) -> str:
-    """Generate a link to unsubscribe from all notifications"""
-    user_hash = user_id.encode("utf-8").decode("unicode_escape")
-    logger.info(f"Generating unsubscribe link for user {user_id}: {user_hash}")
-    return f"https://platform.agpt.co/profile/settings?unsubscribe={user_hash}"
-
-
-async def unsubscribe_user_by_hash(user_hash: str) -> None:
-    """Unsubscribe a user from all notifications"""
-    user_id = user_hash.encode("utf-8").decode("unicode_escape")
-    try:
-        user = await get_user_by_id(user_id)
-        await update_user_notification_preference(
-            user.id,
-            NotificationPreferenceDTO(
-                email=user.email,
-                daily_limit=0,
-                preferences={
-                    NotificationType.AGENT_RUN: False,
-                    NotificationType.ZERO_BALANCE: False,
-                    NotificationType.LOW_BALANCE: False,
-                    NotificationType.BLOCK_EXECUTION_FAILED: False,
-                    NotificationType.CONTINUOUS_AGENT_ERROR: False,
-                    NotificationType.DAILY_SUMMARY: False,
-                    NotificationType.WEEKLY_SUMMARY: False,
-                    NotificationType.MONTHLY_SUMMARY: False,
-                },
-            ),
-        )
-    except Exception as e:
-        raise DatabaseError(
-            f"Failed to unsubscribe user by hash {user_hash}: {e}"
         ) from e
