+++ autogpt_platform/backend/backend/data/user.py
@@ -3,6 +3,7 @@ from datetime import datetime, timedelta
 from typing import Optional, cast
 
 from autogpt_libs.auth.models import DEFAULT_USER_ID
+from backend.util.settings import Settings
 from fastapi import HTTPException
 from prisma import Json
 from prisma.enums import NotificationType
@@ -333,4 +334,42 @@ async def get_user_email_verification(user_id: str) -> bool:
     except Exception as e:
         raise DatabaseError(
             f"Failed to get email verification status for user {user_id}: {e}"
+        ) from e
+
+
+def generate_unsubscribe_link(user_id: str) -> str:
+    """Generate a link to unsubscribe from all notifications"""
+    user_hash = user_id.encode("utf-8").decode("unicode_escape")
+    logger.info(f"Generating unsubscribe link for user {user_id}: {user_hash}")
+
+    settings = Settings()
+    base_url = settings.config.platform_base_url
+    return f"{base_url}/api/email/unsubscribe?user_hash={user_hash}"
+
+
+async def unsubscribe_user_by_hash(user_hash: str) -> None:
+    """Unsubscribe a user from all notifications"""
+    user_id = user_hash.encode("utf-8").decode("unicode_escape")
+    try:
+        user = await get_user_by_id(user_id)
+        await update_user_notification_preference(
+            user.id,
+            NotificationPreferenceDTO(
+                email=user.email,
+                daily_limit=0,
+                preferences={
+                    NotificationType.AGENT_RUN: False,
+                    NotificationType.ZERO_BALANCE: False,
+                    NotificationType.LOW_BALANCE: False,
+                    NotificationType.BLOCK_EXECUTION_FAILED: False,
+                    NotificationType.CONTINUOUS_AGENT_ERROR: False,
+                    NotificationType.DAILY_SUMMARY: False,
+                    NotificationType.WEEKLY_SUMMARY: False,
+                    NotificationType.MONTHLY_SUMMARY: False,
+                },
+            ),
+        )
+    except Exception as e:
+        raise DatabaseError(
+            f"Failed to unsubscribe user by hash {user_hash}: {e}"
         ) from e
